Split out kafka_client_supplier
